feat: move all useHooks in the internal directory to the root of hooks
