+++ src/components/dashboard/widgets/vehicle-status.tsx
@@ -13,8 +13,8 @@ import type { PieSectorDataItem } from "recharts/types/polar/Pie";
 
 import { CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 
-import { useAuthValues } from "@/hooks/internal/useAuthValues";
-import { useTernaryDarkMode } from "@/hooks/internal/useTernaryDarkMode";
+import { useAuthValues } from "@/hooks/useAuthValues";
+import { useTernaryDarkMode } from "@/hooks/useTernaryDarkMode";
 
 import { APP_DEFAULTS, USER_STORAGE_KEYS } from "@/utils/constants";
 import { fetchDashboardVehicleStatusCountsOptions } from "@/utils/query/dashboard";
