+++ src/components/report/page-filters.tsx
@@ -11,7 +11,7 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardTitle } from "@/components/ui/card";
 import { icons } from "@/components/ui/icons";
 
-import { useReportLookupLists } from "@/hooks/internal/useReportLookupLists";
+import { useReportLookupLists } from "@/hooks/useReportLookupLists";
 
 import { useReportContext } from "@/context/view-report";
