+++ src/hooks/useEventListener.ts
@@ -1,6 +1,6 @@
 import { RefObject, useEffect, useRef } from "react";
 
-import { useIsomorphicLayoutEffect } from "@/hooks/internal/useIsomorphicLayoutEffect";
+import { useIsomorphicLayoutEffect } from "@/hooks/useIsomorphicLayoutEffect";
 
 // MediaQueryList Event based useEventListener interface
 function useEventListener<K extends keyof MediaQueryListEventMap>(
