+++ src/hooks/useUpdateEffect.ts
@@ -1,6 +1,6 @@
 import { DependencyList, EffectCallback, useEffect } from "react";
 
-import { useIsFirstRender } from "@/hooks/internal/useIsFirstRender";
+import { useIsFirstRender } from "@/hooks/useIsFirstRender";
 
 export function useUpdateEffect(effect: EffectCallback, deps?: DependencyList) {
   const isFirst = useIsFirstRender();
