+++ src/pages/logged-out.tsx
@@ -1,6 +1,6 @@
 import { Link } from "@tanstack/react-router";
 
-import { useDocumentTitle } from "@/hooks/internal/useDocumentTitle";
+import { useDocumentTitle } from "@/hooks/useDocumentTitle";
 
 import { titleMaker } from "@/utils/title-maker";
