+++ src/pages/search-vehicles.tsx
@@ -18,7 +18,7 @@ import ProtectorShield from "@/components/protector-shield";
 import { Badge } from "@/components/ui/badge";
 import { buttonVariants } from "@/components/ui/button";
 
-import { useDocumentTitle } from "@/hooks/internal/useDocumentTitle";
+import { useDocumentTitle } from "@/hooks/useDocumentTitle";
 import { saveColumnSettings } from "@/api/save-column-settings";
 
 import type { TVehicleListItemParsed } from "@/schemas/vehicle";
